서브태스크 추가

+++ backjoon/src/B15964.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class B15964 {
     public static void main(String[] args) {
         Scanner value = new Scanner(System.in);
-        int a = value.nextInt();
-        int b = value.nextInt();
-        int c = (a+b)*(a-b);
+        long a = value.nextInt();
+        long b = value.nextInt();
+        long c = (a+b)*(a-b);
         System.out.println(c);
     }
 }
